Refactor ucp_management
