+++ cmd/rad/cmd/resourceList.go
@@ -10,6 +10,8 @@ import (
 
 	"github.com/project-radius/radius/pkg/cli"
 	"github.com/project-radius/radius/pkg/cli/environments"
+	"github.com/project-radius/radius/pkg/cli/objectformats"
+	"github.com/project-radius/radius/pkg/cli/output"
 	"github.com/spf13/cobra"
 )
 
@@ -37,11 +39,46 @@ func listResources(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
+	client, err := environments.CreateManagementClient(cmd.Context(), env)
+	if err != nil {
+		return err
+	}
+
+	resourceList, err := client.ListAllResourcesByApplication(cmd.Context(), applicationName)
+	if err != nil {
+		return err
+	}
+
+	format, err := cli.RequireOutput(cmd)
+	if err != nil {
+		return err
+	}
+
+	err = output.Write(format, resourceList.Value, cmd.OutOrStdout(), objectformats.GetResourceTableFormat())
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func listResourcesUCP(cmd *cobra.Command, args []string) error {
+	config := ConfigFromContext(cmd.Context())
+	env, err := cli.RequireEnvironment(cmd, config)
+	if err != nil {
+		return err
+	}
+
+	applicationName, err := cli.RequireApplicationArgs(cmd, args, env)
+	if err != nil {
+		return err
+	}
+
 	client, err := environments.CreateFirstPartyServiceManagementClient(cmd.Context(), env)
 	if err != nil {
 		return err
 	}
-	resourceList, err :=client.ListAllResourcesByApplication(cmd.Context(), applicationName)
+	resourceList, err := client.ListAllResourcesByApplication(cmd.Context(), applicationName)
 	if err != nil {
 		return err
 	}
