+++ pkg/cli/environments/dev.go
@@ -204,7 +204,7 @@ func (e *LocalEnvironment) CreateUCPManagementClient(ctx context.Context) (clien
 		return nil, err
 	}
 
-	return &azure.ARMUCPManagementClient{
+	return &kubernetes.ARMUCPManagementClient {
 		EnvironmentName: e.Name,
 		Connection:      connection,
 		ResourceGroup:   e.Namespace, // Temporarily set resource group and subscription id to the namespace
